Add files via upload
This project provides the customer with a restaurant menu and allows him to browse all sections of the menu with an option to return to the main menu. The project keeps track of all items ordered and provides the customer with a receipt after the order has been placed. The receipt has a list of all items ordered plus the total of the order including tax.