Add files via upload
This program displays the active roster of a football team sorted by position. The program displays team statistics
such as the number of players on the active roster, the current team salary, the teams's cap space, the average age of each player.
The program also includes options to give raises to players, cut players, or add players to the roster.